created util files for better file decomposition
