+++ tensorflow/c/eager/mnist_gradients_test.cc
@@ -13,6 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 #include "tensorflow/c/eager/gradients.h"
+#include "tensorflow/c/eager/mnist_gradients_util.h"
 
 #include <memory>
 
@@ -44,48 +45,48 @@ class CppGradients
 };
 
 // Creates an Identity op.
-Status Identity(AbstractContext* ctx,
-                absl::Span<AbstractTensorHandle* const> inputs,
-                absl::Span<AbstractTensorHandle*> outputs, const char* name) {
+// Status Identity(AbstractContext* ctx,
+//                 absl::Span<AbstractTensorHandle* const> inputs,
+//                 absl::Span<AbstractTensorHandle*> outputs, const char* name) {
   
-  AbstractOperationPtr identity_op(ctx->CreateOperation());
-  TF_RETURN_IF_ERROR(
-      identity_op->Reset("Identity", /*raw_device_name=*/nullptr));
-  if (isa<tracing::TracingOperation>(identity_op.get())) {
-    TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingOperation>(identity_op.get())
-                           ->SetOpName(name));
-  }
-  TF_RETURN_IF_ERROR(identity_op->AddInput(inputs[0]));
-  int num_retvals = 1;
-  TF_RETURN_IF_ERROR(identity_op->Execute(outputs, &num_retvals));
-  return Status::OK();
-}
+//   AbstractOperationPtr identity_op(ctx->CreateOperation());
+//   TF_RETURN_IF_ERROR(
+//       identity_op->Reset("Identity", /*raw_device_name=*/nullptr));
+//   if (isa<tracing::TracingOperation>(identity_op.get())) {
+//     TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingOperation>(identity_op.get())
+//                            ->SetOpName(name));
+//   }
+//   TF_RETURN_IF_ERROR(identity_op->AddInput(inputs[0]));
+//   int num_retvals = 1;
+//   TF_RETURN_IF_ERROR(identity_op->Execute(outputs, &num_retvals));
+//   return Status::OK();
+// }
 
-// Creates a MatMul op.
-Status MatMul(AbstractContext* ctx,
-                absl::Span<AbstractTensorHandle* const> inputs,
-                absl::Span<AbstractTensorHandle*> outputs, const char* name,
-                bool transpose_a, bool transpose_b) {
+// // Creates a MatMul op.
+// Status MatMul(AbstractContext* ctx,
+//                 absl::Span<AbstractTensorHandle* const> inputs,
+//                 absl::Span<AbstractTensorHandle*> outputs, const char* name,
+//                 bool transpose_a, bool transpose_b) {
   
-  AbstractOperationPtr matmul_op(ctx->CreateOperation());
-  TF_RETURN_IF_ERROR(
-      matmul_op->Reset("MatMul", /*raw_device_name=*/nullptr));
+//   AbstractOperationPtr matmul_op(ctx->CreateOperation());
+//   TF_RETURN_IF_ERROR(
+//       matmul_op->Reset("MatMul", /*raw_device_name=*/nullptr));
 
-  if (isa<tracing::TracingOperation>(matmul_op.get())) {
-    TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingOperation>(matmul_op.get())
-                           ->SetOpName(name));
-  }
+//   if (isa<tracing::TracingOperation>(matmul_op.get())) {
+//     TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingOperation>(matmul_op.get())
+//                            ->SetOpName(name));
+//   }
 
-  TF_RETURN_IF_ERROR(matmul_op->AddInput(inputs[0]));
-  TF_RETURN_IF_ERROR(matmul_op->AddInput(inputs[1]));
+//   TF_RETURN_IF_ERROR(matmul_op->AddInput(inputs[0]));
+//   TF_RETURN_IF_ERROR(matmul_op->AddInput(inputs[1]));
 
-  matmul_op->SetAttrBool("transpose_a", transpose_a);
-  matmul_op->SetAttrBool("transpose_b", transpose_b);
+//   matmul_op->SetAttrBool("transpose_a", transpose_a);
+//   matmul_op->SetAttrBool("transpose_b", transpose_b);
   
-  int num_retvals = 1;
-  TF_RETURN_IF_ERROR(matmul_op->Execute(outputs, &num_retvals));
-  return Status::OK();
-}
+//   int num_retvals = 1;
+//   TF_RETURN_IF_ERROR(matmul_op->Execute(outputs, &num_retvals));
+//   return Status::OK();
+// }
 
 
 // =================== Register gradients for Add ============================
@@ -172,142 +173,159 @@ Status RegisterGradientMatMul(GradientRegistry* registry) {
   return registry->Register("MatMul", MatMulRegisterer);
 }
 
+// class ReluGradientFunction : public GradientFunction {
+//  public:
+//   explicit ReluMulGradientFunction(AbstractContext* ctx, std::vector<AbstractTensorHandle*> f_inputs) : 
+//             ctx_(ctx), forward_inputs(f_inputs) {}
+  
+//   Status Compute(absl::Span<AbstractTensorHandle* const> grad_inputs,
+//                  std::vector<AbstractTensorHandle*>* grad_outputs) override {
+    
+//     /* Given upstream grad U and a matmul op A*B, the gradients are:
+//      *      
+//      *    dA = U * B.T   
+//      *    dB = A.T * U
+//      *
+//      *    where A.T means `transpose(A)`
+//      */
+
+//     AbstractTensorHandle* upstream_grad = grad_inputs[0];
+//     grad_outputs->resize(2);
+//     std::vector<AbstractTensorHandle*> matmul_outputs(1);
+
+//     // Gradient for A
+//     TF_RETURN_IF_ERROR(MatMul(ctx_, {upstream_grad, forward_inputs[1]},
+//                               absl::MakeSpan(matmul_outputs), "mm0",  
+//                               /*transpose_a = */false, /*transpose_b = */true));
+
+//     (*grad_outputs)[0] = matmul_outputs[0];
+
+//     // Gradient for B
+//     TF_RETURN_IF_ERROR(MatMul(ctx_, {upstream_grad},
+//                               absl::MakeSpan(matmul_outputs), "mm1", 
+//                               /*transpose_a = */true, /*transpose_b = */false));
+
+//     (*grad_outputs)[1] = matmul_outputs[0];
+//     return Status::OK();
+//   }
+//   ~MatMulGradientFunction() override {}
+
+//  private:
+//   AbstractContext* ctx_;
+//   std::vector<AbstractTensorHandle*> forward_inputs;
+
+// };
+
+// GradientFunction* MatMulRegisterer(const ForwardOperation& op) {
+//   return new MatMulGradientFunction(op.ctx, op.inputs);
+// }
+ 
+// Status RegisterGradientMatMul(GradientRegistry* registry) {
+//   return registry->Register("MatMul", MatMulRegisterer);
+// }
 // =================== End gradient registrations ============================
 
 // Computes `inputs[0] + inputs[1]` and records it on the tape.
-Status Add(AbstractContext* ctx, Tape* tape,
-           absl::Span<AbstractTensorHandle* const> inputs,
-           absl::Span<AbstractTensorHandle*> outputs,
-           const GradientRegistry& registry) {
+// Status Add(AbstractContext* ctx, Tape* tape,
+//            absl::Span<AbstractTensorHandle* const> inputs,
+//            absl::Span<AbstractTensorHandle*> outputs,
+//            const GradientRegistry& registry) {
   
-  AbstractOperationPtr add_op(ctx->CreateOperation());
-  ForwardOperation forward_op;
-  forward_op.ctx = ctx;
-  TF_RETURN_IF_ERROR(
-      Reset(add_op.get(), "Add", /*raw_device_name=*/nullptr, &forward_op));
-  if (isa<tracing::TracingOperation>(add_op.get())) {
-    TF_RETURN_IF_ERROR(
-        dyn_cast<tracing::TracingOperation>(add_op.get())->SetOpName("my_add"));
-  }
-  TF_RETURN_IF_ERROR(AddInput(add_op.get(), inputs[0], &forward_op));
-  TF_RETURN_IF_ERROR(AddInput(add_op.get(), inputs[1], &forward_op));
-  int num_retvals = 1;
-  return Execute(add_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
-                 registry);
-}
+//   AbstractOperationPtr add_op(ctx->CreateOperation());
+//   ForwardOperation forward_op;
+//   forward_op.ctx = ctx;
+//   TF_RETURN_IF_ERROR(
+//       Reset(add_op.get(), "Add", /*raw_device_name=*/nullptr, &forward_op));
+//   if (isa<tracing::TracingOperation>(add_op.get())) {
+//     TF_RETURN_IF_ERROR(
+//         dyn_cast<tracing::TracingOperation>(add_op.get())->SetOpName("my_add"));
+//   }
+//   TF_RETURN_IF_ERROR(AddInput(add_op.get(), inputs[0], &forward_op));
+//   TF_RETURN_IF_ERROR(AddInput(add_op.get(), inputs[1], &forward_op));
+//   int num_retvals = 1;
+//   return Execute(add_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
+//                  registry);
+// }
 
-// Computes `inputs[0] * inputs[1]` for matrices and records it on the tape.
-Status MatMul(AbstractContext* ctx, Tape* tape,
-           absl::Span<AbstractTensorHandle* const> inputs,
-           absl::Span<AbstractTensorHandle*> outputs, const char* name,
-           bool transpose_a, bool transpose_b,
-           const GradientRegistry& registry) {
+// // Computes `inputs[0] * inputs[1]` for matrices and records it on the tape.
+// Status MatMul(AbstractContext* ctx, Tape* tape,
+//            absl::Span<AbstractTensorHandle* const> inputs,
+//            absl::Span<AbstractTensorHandle*> outputs, const char* name,
+//            bool transpose_a, bool transpose_b,
+//            const GradientRegistry& registry) {
   
-  AbstractOperationPtr matmul_op(ctx->CreateOperation());
-  ForwardOperation forward_op;
-  forward_op.ctx = ctx;
-  TF_RETURN_IF_ERROR(
-      Reset(matmul_op.get(), "MatMul", /*raw_device_name=*/nullptr, &forward_op));
-  if (isa<tracing::TracingOperation>(matmul_op.get())) {
-    TF_RETURN_IF_ERROR(
-        dyn_cast<tracing::TracingOperation>(matmul_op.get())->SetOpName(name));
-  }
+//   AbstractOperationPtr matmul_op(ctx->CreateOperation());
+//   ForwardOperation forward_op;
+//   forward_op.ctx = ctx;
+//   TF_RETURN_IF_ERROR(
+//       Reset(matmul_op.get(), "MatMul", /*raw_device_name=*/nullptr, &forward_op));
+//   if (isa<tracing::TracingOperation>(matmul_op.get())) {
+//     TF_RETURN_IF_ERROR(
+//         dyn_cast<tracing::TracingOperation>(matmul_op.get())->SetOpName(name));
+//   }
 
-  TF_RETURN_IF_ERROR(AddInput(matmul_op.get(), inputs[0], &forward_op));
-  TF_RETURN_IF_ERROR(AddInput(matmul_op.get(), inputs[1], &forward_op));
-  matmul_op->SetAttrBool("transpose_a", transpose_a);
-  matmul_op->SetAttrBool("transpose_b", transpose_b);
+//   TF_RETURN_IF_ERROR(AddInput(matmul_op.get(), inputs[0], &forward_op));
+//   TF_RETURN_IF_ERROR(AddInput(matmul_op.get(), inputs[1], &forward_op));
+//   matmul_op->SetAttrBool("transpose_a", transpose_a);
+//   matmul_op->SetAttrBool("transpose_b", transpose_b);
 
-  int num_retvals = 1;
-  return Execute(matmul_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
-                 registry);
-}
+//   int num_retvals = 1;
+//   return Execute(matmul_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
+//                  registry);
+// }
 
-// Computes `Relu(inputs[0])` and records it on the tape.
-Status Relu(AbstractContext* ctx, Tape* tape,
-           absl::Span<AbstractTensorHandle* const> inputs,
-           absl::Span<AbstractTensorHandle*> outputs, const char* name,
-           const GradientRegistry& registry) {
+// // Computes `Relu(inputs[0])` and records it on the tape.
+// Status Relu(AbstractContext* ctx, Tape* tape,
+//            absl::Span<AbstractTensorHandle* const> inputs,
+//            absl::Span<AbstractTensorHandle*> outputs, const char* name,
+//            const GradientRegistry& registry) {
   
-  AbstractOperationPtr relu_op(ctx->CreateOperation());
-  ForwardOperation forward_op;
-  forward_op.ctx = ctx;
-  TF_RETURN_IF_ERROR(
-      Reset(relu_op.get(), "Relu", /*raw_device_name=*/nullptr, &forward_op));
-  if (isa<tracing::TracingOperation>(relu_op.get())) {
-    TF_RETURN_IF_ERROR(
-        dyn_cast<tracing::TracingOperation>(relu_op.get())->SetOpName(name));
-  }
-  TF_RETURN_IF_ERROR(AddInput(relu_op.get(), inputs[0], &forward_op));
-  int num_retvals = 1;
-  return Execute(relu_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
-                 registry);
-}
+//   AbstractOperationPtr relu_op(ctx->CreateOperation());
+//   ForwardOperation forward_op;
+//   forward_op.ctx = ctx;
+//   TF_RETURN_IF_ERROR(
+//       Reset(relu_op.get(), "Relu", /*raw_device_name=*/nullptr, &forward_op));
+//   if (isa<tracing::TracingOperation>(relu_op.get())) {
+//     TF_RETURN_IF_ERROR(
+//         dyn_cast<tracing::TracingOperation>(relu_op.get())->SetOpName(name));
+//   }
+//   TF_RETURN_IF_ERROR(AddInput(relu_op.get(), inputs[0], &forward_op));
+//   int num_retvals = 1;
+//   return Execute(relu_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
+//                  registry);
+// }
 
-// Computes `SoftmaxLoss(scores, labels)` for matrices and records it on the tape.
-Status SparseSoftmaxCrossEntropyLoss(AbstractContext* ctx, Tape* tape,
-           absl::Span<AbstractTensorHandle* const> inputs,
-           absl::Span<AbstractTensorHandle*> outputs, const char* name,
-           const GradientRegistry& registry) {
+// // Computes `SoftmaxLoss(scores, labels)` and records it on the tape.
+// Status SparseSoftmaxCrossEntropyLoss(AbstractContext* ctx, Tape* tape,
+//            absl::Span<AbstractTensorHandle* const> inputs,
+//            absl::Span<AbstractTensorHandle*> outputs, const char* name,
+//            const GradientRegistry& registry) {
   
-  AbstractTensorHandle* scores = inputs[0];
-  AbstractTensorHandle* labels = inputs[1];
+//   AbstractTensorHandle* scores = inputs[0];
+//   AbstractTensorHandle* labels = inputs[1];
 
-  AbstractOperationPtr sm_op(ctx->CreateOperation());
-  ForwardOperation forward_op;
-  forward_op.ctx = ctx;
-  TF_RETURN_IF_ERROR(
-      Reset(sm_op.get(), "SparseSoftmaxCrossEntropyWithLogits", /*raw_device_name=*/nullptr, &forward_op));
-  if (isa<tracing::TracingOperation>(sm_op.get())) {
-    TF_RETURN_IF_ERROR(
-        dyn_cast<tracing::TracingOperation>(sm_op.get())->SetOpName(name));
-  }
+//   AbstractOperationPtr sm_op(ctx->CreateOperation());
+//   ForwardOperation forward_op;
+//   forward_op.ctx = ctx;
+//   TF_RETURN_IF_ERROR(
+//       Reset(sm_op.get(), "SparseSoftmaxCrossEntropyWithLogits", /*raw_device_name=*/nullptr, &forward_op));
+//   if (isa<tracing::TracingOperation>(sm_op.get())) {
+//     TF_RETURN_IF_ERROR(
+//         dyn_cast<tracing::TracingOperation>(sm_op.get())->SetOpName(name));
+//   }
 
-  TF_RETURN_IF_ERROR(AddInput(sm_op.get(), scores, &forward_op));
-  TF_RETURN_IF_ERROR(AddInput(sm_op.get(), labels, &forward_op));
+//   TF_RETURN_IF_ERROR(AddInput(sm_op.get(), scores, &forward_op));
+//   TF_RETURN_IF_ERROR(AddInput(sm_op.get(), labels, &forward_op));
 
-  int num_retvals = 2; // returns loss values and backprop
-  return Execute(sm_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
-                 registry);
-}
-
-// Computes
-// y = inputs[0] + inputs[1]
-// return grad(y, {inputs[0], inputs[1]})
-Status AddGradModel(AbstractContext* ctx,
-                    absl::Span<AbstractTensorHandle* const> inputs,
-                    absl::Span<AbstractTensorHandle*> outputs,
-                    const GradientRegistry& registry) {
-  
-  TapeVSpace vspace(ctx);
-  auto tape = new Tape(/*persistent=*/false);
-  tape->Watch(ToId(inputs[0]));  // Watch x.
-  tape->Watch(ToId(inputs[1]));  // Watch y.
-  std::vector<AbstractTensorHandle*> add_outputs(1);
-  TF_RETURN_IF_ERROR(Add(ctx, tape, inputs, absl::MakeSpan(add_outputs),
-                         registry));  // Compute x+y.
-  std::unordered_map<tensorflow::int64, TapeTensor>
-      source_tensors_that_are_targets;
-
-  std::vector<AbstractTensorHandle*> out_grads;
-  TF_RETURN_IF_ERROR(tape->ComputeGradient(
-      vspace, /*target_tensor_ids=*/{ToId(add_outputs[0])},
-      /*source_tensor_ids=*/{ToId(inputs[0]), ToId(inputs[1])},
-      source_tensors_that_are_targets,
-      /*output_gradients=*/{}, &out_grads));
-  for (auto add_output : add_outputs) {
-    add_output->Release();
-  }
-  outputs[0] = out_grads[0];
-  outputs[1] = out_grads[1];
-  delete tape;
-  return Status::OK();
-}
+//   int num_retvals = 2; // returns loss values and backprop
+//   return Execute(sm_op.get(), ctx, outputs, &num_retvals, &forward_op, tape,
+//                  registry);
+// }
 
 // // Computes
-// // y = inputs[0] * inputs[1]
+// // y = inputs[0] + inputs[1]
 // // return grad(y, {inputs[0], inputs[1]})
-// Status MatMulGradModel(AbstractContext* ctx,
+// Status AddGradModel(AbstractContext* ctx,
 //                     absl::Span<AbstractTensorHandle* const> inputs,
 //                     absl::Span<AbstractTensorHandle*> outputs,
 //                     const GradientRegistry& registry) {
@@ -316,21 +334,20 @@ Status AddGradModel(AbstractContext* ctx,
 //   auto tape = new Tape(/*persistent=*/false);
 //   tape->Watch(ToId(inputs[0]));  // Watch x.
 //   tape->Watch(ToId(inputs[1]));  // Watch y.
-//   std::vector<AbstractTensorHandle*> mm_outputs(1);
-//   TF_RETURN_IF_ERROR(MatMul(ctx, tape, inputs, absl::MakeSpan(mm_outputs), 
-//       "matmul0", /*transpose_a=*/false, /*transpose_b=*/false, registry));  // Compute x*y.
-  
+//   std::vector<AbstractTensorHandle*> add_outputs(1);
+//   TF_RETURN_IF_ERROR(Add(ctx, tape, inputs, absl::MakeSpan(add_outputs),
+//                          registry));  // Compute x+y.
 //   std::unordered_map<tensorflow::int64, TapeTensor>
 //       source_tensors_that_are_targets;
 
 //   std::vector<AbstractTensorHandle*> out_grads;
 //   TF_RETURN_IF_ERROR(tape->ComputeGradient(
-//       vspace, /*target_tensor_ids=*/{ToId(mm_outputs[0])},
+//       vspace, /*target_tensor_ids=*/{ToId(add_outputs[0])},
 //       /*source_tensor_ids=*/{ToId(inputs[0]), ToId(inputs[1])},
 //       source_tensors_that_are_targets,
 //       /*output_gradients=*/{}, &out_grads));
-//   for (auto mm_output : mm_outputs) {
-//     mm_output->Release();
+//   for (auto add_output : add_outputs) {
+//     add_output->Release();
 //   }
 //   outputs[0] = out_grads[0];
 //   outputs[1] = out_grads[1];
@@ -338,126 +355,126 @@ Status AddGradModel(AbstractContext* ctx,
 //   return Status::OK();
 // }
 
-Status MNISTForwardModel(AbstractContext* ctx,
-                    absl::Span<AbstractTensorHandle* const> inputs,
-                    absl::Span<AbstractTensorHandle*> outputs,
-                    const GradientRegistry& registry) {
-  /**
-    * We will trace a 2-layer fully connected network for an MNIST model:
-    *
-    *   def mnist_forward(X, W1, W2, y_labels):
-    *     mm_out_1 = tf.matmul(X,W1)
-    *     hidden_layer = tf.ReLu(mm_out_1)
-    *     scores = tf.matmul(hidden_layer,W2)
-    *     softmax = tf.softmaxLoss(scores,y_labels)
-    *     return scores, softmax 
-    *
-    * Use this convention for inputs:
-    * 
-    *   inputs = [X, W1, W2, y_labels]
-    *
-    */ 
-  AbstractTensorHandle* X = inputs[0];
-  AbstractTensorHandle* W1 = inputs[1];
-  AbstractTensorHandle* W2 = inputs[2];
-  AbstractTensorHandle* y_labels = inputs[3];
+// Status MNISTForwardModel(AbstractContext* ctx,
+//                     absl::Span<AbstractTensorHandle* const> inputs,
+//                     absl::Span<AbstractTensorHandle*> outputs,
+//                     const GradientRegistry& registry) {
+//   /**
+//     * We will trace a 2-layer fully connected network for an MNIST model:
+//     *
+//     *   def mnist_forward(X, W1, W2, y_labels):
+//     *     mm_out_1 = tf.matmul(X,W1)
+//     *     hidden_layer = tf.ReLu(mm_out_1)
+//     *     scores = tf.matmul(hidden_layer,W2)
+//     *     softmax = tf.softmaxLoss(scores,y_labels)
+//     *     return scores, softmax 
+//     *
+//     * Use this convention for inputs:
+//     * 
+//     *   inputs = [X, W1, W2, y_labels]
+//     *
+//     */ 
+//   AbstractTensorHandle* X = inputs[0];
+//   AbstractTensorHandle* W1 = inputs[1];
+//   AbstractTensorHandle* W2 = inputs[2];
+//   AbstractTensorHandle* y_labels = inputs[3];
 
-  TapeVSpace vspace(ctx);
-  auto tape = new Tape(/*persistent=*/false);
-  tape->Watch(ToId(W1));  // Watch W1.
-  tape->Watch(ToId(W2));  // Watch W2.
-  std::vector<AbstractTensorHandle*> temp_outputs(1);
+//   TapeVSpace vspace(ctx);
+//   auto tape = new Tape(/*persistent=*/false);
+//   tape->Watch(ToId(W1));  // Watch W1.
+//   tape->Watch(ToId(W2));  // Watch W2.
+//   std::vector<AbstractTensorHandle*> temp_outputs(1);
 
-  TF_RETURN_IF_ERROR(MatMul(ctx, tape, {X, W1}, absl::MakeSpan(temp_outputs),
-                     "matmul0",/*transpose_a=*/false,/*transpose_b=*/false, registry));  // Compute X*W1
+//   TF_RETURN_IF_ERROR(MatMul(ctx, tape, {X, W1}, absl::MakeSpan(temp_outputs),
+//                      "matmul0",/*transpose_a=*/false,/*transpose_b=*/false, registry));  // Compute X*W1
 
-  TF_RETURN_IF_ERROR(Relu(ctx, tape, {temp_outputs[0]}, absl::MakeSpan(temp_outputs),
-                     "relu", registry)); // Compute Relu(X*W1)
+//   TF_RETURN_IF_ERROR(Relu(ctx, tape, {temp_outputs[0]}, absl::MakeSpan(temp_outputs),
+//                      "relu", registry)); // Compute Relu(X*W1)
 
-  TF_RETURN_IF_ERROR(MatMul(ctx, tape, {temp_outputs[0], W2}, absl::MakeSpan(temp_outputs),
-                     "matmul1",/*transpose_a=*/false,/*transpose_b=*/false, registry));  // Compute W2*Relu(X*W1)
+//   TF_RETURN_IF_ERROR(MatMul(ctx, tape, {temp_outputs[0], W2}, absl::MakeSpan(temp_outputs),
+//                      "matmul1",/*transpose_a=*/false,/*transpose_b=*/false, registry));  // Compute W2*Relu(X*W1)
   
-  AbstractTensorHandle* scores = temp_outputs[0];
+//   AbstractTensorHandle* scores = temp_outputs[0];
 
-  TF_RETURN_IF_ERROR(SparseSoftmaxCrossEntropyLoss(ctx, tape, {scores, y_labels}, absl::MakeSpan(temp_outputs),
-                     "softmax_loss", registry));  // Compute Softmax(Scores,labels)
+//   TF_RETURN_IF_ERROR(SparseSoftmaxCrossEntropyLoss(ctx, tape, {scores, y_labels}, absl::MakeSpan(temp_outputs),
+//                      "softmax_loss", registry));  // Compute Softmax(Scores,labels)
   
-  AbstractTensorHandle* loss_vals  = temp_outputs[0];
-  outputs[0] = scores;
-  outputs[1] = loss_vals;
-  delete tape;
-  return Status::OK();
-}
+//   AbstractTensorHandle* loss_vals  = temp_outputs[0];
+//   outputs[0] = scores;
+//   outputs[1] = loss_vals;
+//   delete tape;
+//   return Status::OK();
+// }
 
-AbstractContext* BuildFunction(const char* fn_name) {
-  std::unique_ptr<TF_Status, decltype(&TF_DeleteStatus)> status(
-      TF_NewStatus(), TF_DeleteStatus);
-  TF_ExecutionContext* graph_ctx = TF_CreateFunction(fn_name, status.get());
-  return unwrap(graph_ctx);
-}
+// AbstractContext* BuildFunction(const char* fn_name) {
+//   std::unique_ptr<TF_Status, decltype(&TF_DeleteStatus)> status(
+//       TF_NewStatus(), TF_DeleteStatus);
+//   TF_ExecutionContext* graph_ctx = TF_CreateFunction(fn_name, status.get());
+//   return unwrap(graph_ctx);
+// }
 
-Status CreateParamsForInputs(AbstractContext* ctx,
-                             absl::Span<AbstractTensorHandle* const> inputs,
-                             std::vector<AbstractTensorHandle*>* params) {
+// Status CreateParamsForInputs(AbstractContext* ctx,
+//                              absl::Span<AbstractTensorHandle* const> inputs,
+//                              std::vector<AbstractTensorHandle*>* params) {
   
-  tracing::TracingTensorHandle* handle = nullptr;
-  for (auto input : inputs) {
-    TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingContext>(ctx)->AddParameter(
-        input->DataType(), &handle));
-    params->emplace_back(handle);
-  }
-  return Status::OK();
-}
+//   tracing::TracingTensorHandle* handle = nullptr;
+//   for (auto input : inputs) {
+//     TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingContext>(ctx)->AddParameter(
+//         input->DataType(), &handle));
+//     params->emplace_back(handle);
+//   }
+//   return Status::OK();
+// }
 
-using Model = std::function<Status(
-    AbstractContext*, absl::Span<AbstractTensorHandle* const>,
-    absl::Span<AbstractTensorHandle*>, const GradientRegistry&)>;
+// using Model = std::function<Status(
+//     AbstractContext*, absl::Span<AbstractTensorHandle* const>,
+//     absl::Span<AbstractTensorHandle*>, const GradientRegistry&)>;
 
-// Runs `model` maybe wrapped in a function.
-Status RunModel(Model model, AbstractContext* ctx,
-                absl::Span<AbstractTensorHandle* const> inputs,
-                absl::Span<AbstractTensorHandle*> outputs, bool use_function,
-                const GradientRegistry& registry) {
+// // Runs `model` maybe wrapped in a function.
+// Status RunModel(Model model, AbstractContext* ctx,
+//                 absl::Span<AbstractTensorHandle* const> inputs,
+//                 absl::Span<AbstractTensorHandle*> outputs, bool use_function,
+//                 const GradientRegistry& registry) {
   
-  if (use_function) {
-    const char* fn_name = "test_fn";
-    std::unique_ptr<AbstractFunction> scoped_func;
-    {
-      AbstractContextPtr func_ctx(BuildFunction(fn_name));
-      std::vector<AbstractTensorHandle*> func_inputs;
-      func_inputs.reserve(inputs.size());
-      TF_RETURN_IF_ERROR(
-          CreateParamsForInputs(func_ctx.get(), inputs, &func_inputs));
-      OutputList output_list;
-      output_list.expected_num_outputs = outputs.size();
-      output_list.outputs.resize(outputs.size());
-      TF_RETURN_IF_ERROR(model(func_ctx.get(), absl::MakeSpan(func_inputs),
-                               absl::MakeSpan(output_list.outputs), registry));
-      for (auto func_input : func_inputs) {
-        func_input->Release();
-      }
-      AbstractFunction* func = nullptr;
-      TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingContext>(func_ctx.get())
-                             ->Finalize(&output_list, &func));
-      scoped_func.reset(func);
-      output_list.outputs[0]->Release();
-      output_list.outputs[1]->Release();
-      TF_RETURN_IF_ERROR(ctx->RegisterFunction(func));
-    }
+//   if (use_function) {
+//     const char* fn_name = "test_fn";
+//     std::unique_ptr<AbstractFunction> scoped_func;
+//     {
+//       AbstractContextPtr func_ctx(BuildFunction(fn_name));
+//       std::vector<AbstractTensorHandle*> func_inputs;
+//       func_inputs.reserve(inputs.size());
+//       TF_RETURN_IF_ERROR(
+//           CreateParamsForInputs(func_ctx.get(), inputs, &func_inputs));
+//       OutputList output_list;
+//       output_list.expected_num_outputs = outputs.size();
+//       output_list.outputs.resize(outputs.size());
+//       TF_RETURN_IF_ERROR(model(func_ctx.get(), absl::MakeSpan(func_inputs),
+//                                absl::MakeSpan(output_list.outputs), registry));
+//       for (auto func_input : func_inputs) {
+//         func_input->Release();
+//       }
+//       AbstractFunction* func = nullptr;
+//       TF_RETURN_IF_ERROR(dyn_cast<tracing::TracingContext>(func_ctx.get())
+//                              ->Finalize(&output_list, &func));
+//       scoped_func.reset(func);
+//       output_list.outputs[0]->Release();
+//       output_list.outputs[1]->Release();
+//       TF_RETURN_IF_ERROR(ctx->RegisterFunction(func));
+//     }
 
-    AbstractOperationPtr fn_op(ctx->CreateOperation());
-    TF_RETURN_IF_ERROR(fn_op->Reset(fn_name, /*raw_device_name=*/nullptr));
-    for (auto input : inputs) {
-      TF_RETURN_IF_ERROR(fn_op->AddInput(input));
-    }
-    int retvals = outputs.size();
-    TF_RETURN_IF_ERROR(fn_op->Execute(outputs, &retvals));
-    TF_RETURN_IF_ERROR(ctx->RemoveFunction(fn_name));
-    return Status::OK();
-  } else {
-    return model(ctx, inputs, outputs, registry);
-  }
-}
+//     AbstractOperationPtr fn_op(ctx->CreateOperation());
+//     TF_RETURN_IF_ERROR(fn_op->Reset(fn_name, /*raw_device_name=*/nullptr));
+//     for (auto input : inputs) {
+//       TF_RETURN_IF_ERROR(fn_op->AddInput(input));
+//     }
+//     int retvals = outputs.size();
+//     TF_RETURN_IF_ERROR(fn_op->Execute(outputs, &retvals));
+//     TF_RETURN_IF_ERROR(ctx->RemoveFunction(fn_name));
+//     return Status::OK();
+//   } else {
+//     return model(ctx, inputs, outputs, registry);
+//   }
+// }
 
 void printArr(float data[], int n)
 {
@@ -471,16 +488,16 @@ void printArr(float data[], int n)
 }
 
 
-Status BuildImmediateExecutionContext(bool use_tfrt, AbstractContext** ctx) {
-  std::unique_ptr<TF_Status, decltype(&TF_DeleteStatus)> status(
-      TF_NewStatus(), TF_DeleteStatus);
-  TFE_ContextOptions* opts = TFE_NewContextOptions();
-  TFE_ContextOptionsSetTfrt(opts, use_tfrt);
-  *ctx = unwrap(TF_NewEagerExecutionContext(opts, status.get()));
-  TF_RETURN_IF_ERROR(StatusFromTF_Status(status.get()));
-  TFE_DeleteContextOptions(opts);
-  return Status::OK();
-}
+// Status BuildImmediateExecutionContext(bool use_tfrt, AbstractContext** ctx) {
+//   std::unique_ptr<TF_Status, decltype(&TF_DeleteStatus)> status(
+//       TF_NewStatus(), TF_DeleteStatus);
+//   TFE_ContextOptions* opts = TFE_NewContextOptions();
+//   TFE_ContextOptionsSetTfrt(opts, use_tfrt);
+//   *ctx = unwrap(TF_NewEagerExecutionContext(opts, status.get()));
+//   TF_RETURN_IF_ERROR(StatusFromTF_Status(status.get()));
+//   TFE_DeleteContextOptions(opts);
+//   return Status::OK();
+// }
 
 
 // Get a scalar TensorHandle woth given value
@@ -690,6 +707,9 @@ TEST_P(CppGradients, TestMatMulGrad) {
                /*use_function=*/!std::get<2>(GetParam()), registry);
   ASSERT_EQ(errors::OK, s.code()) << s.error_message();
 
+  // s = MatMulGradModel(ctx.get(), {A.get(), B.get()}, absl::MakeSpan(outputs), registry);
+  // ASSERT_EQ(errors::OK, s.code()) << s.error_message();
+
   TF_Tensor* dA_tensor;
   s = getValue(outputs[0], &dA_tensor);
   ASSERT_EQ(errors::OK, s.code()) << s.error_message();
@@ -698,10 +718,10 @@ TEST_P(CppGradients, TestMatMulGrad) {
   memcpy(&result_data[0], TF_TensorData(dA_tensor), TF_TensorByteSize(dA_tensor));
   
   float expected_dA [4] =  {-.5f, 2.0f, -.5f, 2.0f}; 
-  // float tolerance = 1e-3;
-  // for(int j = 0; j < 4; j++){
-  //   ASSERT_NEAR(result_data[j], expected_dA[j], tolerance);
-  // }  
+  float tolerance = 1e-3;
+  for(int j = 0; j < 4; j++){
+    ASSERT_NEAR(result_data[j], expected_dA[j], tolerance);
+  }  
 
 
   /* ERROR: This test runs 2x when we bazel test
@@ -939,13 +959,13 @@ INSTANTIATE_TEST_SUITE_P(
     UnifiedCAPI, CppGradients,
     ::testing::Combine(::testing::Values("graphdef"),
                        /*tfrt*/ ::testing::Values(false),
-                       /*executing_eagerly*/ ::testing::Values(true, false)));
+                       /*executing_eagerly*/ ::testing::Values(true)));  // change back to (true,false)
 #else
 INSTANTIATE_TEST_SUITE_P(
     UnifiedCAPI, CppGradients,
     ::testing::Combine(::testing::Values("graphdef"),
                        /*tfrt*/ ::testing::Values(false),
-                       /*executing_eagerly*/ ::testing::Values(true, false)));
+                       /*executing_eagerly*/ ::testing::Values(true))); // change back to (true,false)
 #endif
 }  // namespace
 }  // namespace internal
